Create Code for WordCount with the use of Python and MapReduce WordCount with Python (Conventional Method)
This section consists of 2 codes. The first code is used for doing the Python Wordcount without the use of MapReduce. The second code is used for doing the Python Wordcount with the use of MapReduce. However, both these codes is done using the conventional method which is done through Python running on our own local disk and not through Big Data Framework (Hadoop).